pyproject.toml
[tool.poetry]
name = "akm_tools"
version = "0.1.0"
description = ""
authors = ["Mohit Yadav <mohit.yadav.id.code.repos@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "*"
jsonschema = "*"
black = "^24.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
